feat: Add Demo2 method to BasicHtmlBuilding class
This commit adds a new method, Demo2, to the BasicHtmlBuilding class. The Demo2 method generates an HTML document with various elements and styles. It also includes a table and displays a list of processes from the system. The openInBrowser parameter allows the user to specify whether the generated HTML document should be opened in a browser. This method enhances the functionality of the BasicHtmlBuilding class and provides more options for generating HTML documents.