feat: Add Direction enum for representing text direction
This commit adds the Direction enum to the HtmlForgeX.Enums namespace. The enum includes two values, Ltr (Left-to-right) and Rtl (Right-to-left), which represent the direction of text in an HTML document. This addition enhances the functionality of the HtmlForgeX library by providing a convenient way to specify text direction in generated HTML documents.