feat: Add Id property and StyleList to DataTablesTable
This commit adds an `Id` property and a `StyleList` property to the `DataTablesTable` class. The `Id` property is used to set the ID attribute of the table, while the `StyleList` property is a list of `BootStrapTableStyle` enum values that determine the styling of the table. This allows for more customization and flexibility when creating DataTables tables.